+++ changelog.md
@@ -1,7 +1,10 @@
 ### Revision History
-* 5.0.0-SNAPSHOT
-  * Replaced `GuavaCache` with `CaffeineCache` to adhere to Spring Framework 2.0 recommendations.
-  * Updated dependencies to utilize Spring Framework 2.0.
+* 4.4.0-SNAPSHOT
+  * Updated to consume `spring-boot 2.0.0.RELEASE`
+* 4.3.1
+  * Regexes updated to better locate NCube names in APIs that search for them in source code.  The `use()` method is now searched.  Nested calls are not searched.  Extract nested call to local variable, and then use that value in 2nd call. Relevant only if you care about searching sources for NCube names.
+  * Updated to consume `groovy 2.4.15`
+  * Updated to consume `spring-boot 1.5.10.RELEASE`
 * 4.3.0
   * `NCubeRuntime` now has the option to cache .json files locally of `NCubes` that have been loaded. Subsequent requests will potentially load from the cache, following the options listed below.  If no `application.properties` entry exists for this, nothing is cached.  Otherwise set the key `ncube.cache.snapshotPolicy` equal to one of the values from `SnapshotPolicy` `enum` and set the key `ncube.cache.dir` to the location of where the files should be stored. Valid cache policy options:
      * `RELEASE_ONLY`: prevent caching of snapshots; only cache release versions (default)